Refactor callback endpoint
