+++ backend/src/api/spotifyAuth.ts
@@ -6,6 +6,7 @@ import { generateRandomString } from '../utility/fileUtils.js';
 
 // Read env variables
 const client_id = process.env.CLIENT_ID || '';
+const client_secret = process.env.CLIENT_SECRET || '';
 
 export async function clientCredentialsFlow(client_id: string, client_secret: string): Promise<[string, string]> {
 
@@ -67,6 +68,36 @@ export function generateOAuthQuerystring(): string {
         redirect_uri: redirectURI,
         state: state
     });
+}
+
+export async function oAuthAuthorization(code: string): Promise<string[]> {
+    const url = 'https://accounts.spotify.com/api/token';
+    const data = {
+        code: code,
+        redirect_uri: 'http://127.0.0.1:3000/callback',
+        grant_type: 'authorization_code'
+    };
+    const config = {
+        headers: {
+            'content-type': 'application/x-www-form-urlencoded',
+            'Authorization': 'Basic ' + (Buffer.from(client_id + ':' + client_secret).toString('base64'))
+        }
+    };
+
+    const response = await axios.post<SpotifyAuthTokenResponse>(url, data, config);
+
+    let access_token = response.data.access_token;
+    let expires_in = response.data.expires_in;
+    let refresh_token = response.data.refresh_token;
+
+    let validUntil: number = Date.now() + (expires_in * 1000);
+
+    process.env.AUTH_CREDENTIAL_TOKEN = access_token;
+    process.env.AUTH_CREDENTIAL_TOKEN_EXPIRATION = String(validUntil);
+    process.env.AUTH_REFRESH_TOKEN = refresh_token;
+
+    return [access_token, String(expires_in), refresh_token];
+
 }
 
 export async function refreshAuthToken() {
